Migrate to VueJS from Backbone
