+++ client/galaxy/scripts/mvc/tool/tool-form.js
@@ -13,7 +13,7 @@ import ToolFormBase from "mvc/tool/tool-form-base";
 import Webhooks from "mvc/webhooks";
 import Vue from "vue";
 import ToolEntryPoints from "components/ToolEntryPoints/ToolEntryPoints";
-import ToolRecommendation from "mvc/tool-recommendation";
+import ToolRecommendation from "components/ToolRecommendation";
 
 const View = Backbone.View.extend({
     initialize: function(options) {
@@ -271,10 +271,15 @@ const View = Backbone.View.extend({
                 const enable_tool_recommendations = window.Galaxy.config.enable_tool_recommendations;
                 if (enable_tool_recommendations === true || enable_tool_recommendations === "true") {
                     // show tool recommendations
-                    this.$el.append($("<div/>", { id: "tool-recommendation-view" }));
-                    new ToolRecommendation.ToolRecommendationView({
-                        toolId: job_def.tool_id
-                    });
+                    const ToolRecommendationInstance = Vue.extend(ToolRecommendation);
+                        const vm = document.createElement("div");
+                        this.$el.append(vm);
+                        const instance = new ToolRecommendationInstance({
+                            propsData: {
+                                toolId: job_def.tool_id
+                            }
+                        });
+                        instance.$mount(vm);
                 }
                 this.$el.parent().scrollTop(0);
                 // Show Webhook if job is running
